projects now renders from array, working on spacing next
